% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGC.R
\name{readGC}
\alias{readGC}
\title{readGC}
\usage{
readGC(file, ncompounds)
}
\arguments{
\item{file}{A text file, exported from the Shimadzu GCsolution or GCMS solution. 
Only the "Compound Quantitative Result" must be clicked when exporting.}

\item{ncompounds}{Integer. Number of compounds, as defined in the compound table. Future versions will 
detect this automatically.}
}
\value{
A tibble containing all the information from the exported file.
The \code{sample} variable contains sample names, generated from the file names 
(matches word/dot/dash/space from the end of file name and removes the .gcd or .qgd extension).
}
\description{
Parser for Shimadzu GCsolution and GCMSsolution data
}
\author{
Ange Angelov
}
