% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGC.R
\name{readGC}
\alias{readGC}
\title{readGC}
\usage{
readGC(file, writexl = FALSE)
}
\arguments{
\item{file}{a text file, exported from the Shimadzu GCsolution or GCMS solution. 
Only the "Compound Quantitative Result" must be clicked when exporting.}

\item{writexl}{logical: write the resulting tibble to xlsx?}
}
\value{
A tibble containing all the information from the exported file.
The \code{sample} variable contains sample names, generated from the file names 
(matches word/dot/dash/space from the end of file name and removes the .gcd or .qgd extension). 
The function tries to detect the number of samples and compounds and prints this information. If the \code{writexl} 
argument is \code{TRUE}, then an excel file with the data is written to the current directory.
}
\description{
Parser for Shimadzu GCsolution and GCMSsolution data
}
\author{
Angel Angelov
}
